Complete Knight Travails Project
This is the BFS version of my project which works properly.

This solution works semi efficiently with a stored memory to not repeat
past moves. The major pitval of it is how verbose the if statements are,
they can be fixed with some simple code, but I think that it would
actually make that section less readable if I had done that.

If this this project is ever expanded to support more tiles the above
stated issue will absolutely have to be solved.